#19 - Finish Flask-SQLAlchemy integration
