+++ db.py
@@ -1,304 +1,363 @@
-import sqlite3
+# Rewrite of db.py to use SQLAlchemy
+import sqlalchemy
+from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy import Integer, String, func, extract
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from typing import Optional
 import datetime
-import api
 
 
-def create_connection(db_file):
-    connection = None
+class Base(DeclarativeBase):
+    id: Mapped[int] = mapped_column(primary_key=True)
+
+
+db = SQLAlchemy(model_class=Base)
+
+
+# --- Database models ---
+class Release(db.Model):
+    __tablename__ = "release"
+    mbid: Mapped[Optional[str]] = mapped_column(String, unique=True)
+    artist_id: Mapped[int] = mapped_column(sqlalchemy.ForeignKey("artist.id"))
+    label_id: Mapped[int] = mapped_column(sqlalchemy.ForeignKey("label.id"))
+    title: Mapped[str] = mapped_column(String())
+    release_year: Mapped[int] = mapped_column(Integer())
+    runtime: Mapped[int] = mapped_column(Integer())
+    rating: Mapped[int] = mapped_column(Integer())
+    listen_date: Mapped[str] = mapped_column(String())
+    track_count: Mapped[int] = mapped_column(Integer())
+    country: Mapped[str] = mapped_column(String())
+    genre: Mapped[str] = mapped_column(String())
+    tags: Mapped[Optional[str]] = mapped_column(String())
+    art: Mapped[Optional[str]] = mapped_column(String())
+
+    def __init__(self, mbid: Optional[str] = None, artist_id: int = 0, label_id: int = 0, **kwargs):
+        self.mbid = mbid
+        self.artist_id = artist_id
+        self.label_id = label_id
+        for key, value in kwargs.items():
+            setattr(self, key, value)
+
+
+class ArtistOrLabel(db.Model):
+    # Artist and Label tables are both built from this prototype
+    __abstract__ = True
+    mbid: Mapped[Optional[str]] = mapped_column(String(), unique=True)
+    name: Mapped[str] = mapped_column(String())
+    country: Mapped[Optional[str]] = mapped_column(String())
+    type: Mapped[Optional[str]] = mapped_column(String())
+    begin_date: Mapped[Optional[str]] = mapped_column(String())
+    end_date: Mapped[Optional[str]] = mapped_column(String())
+    image: Mapped[Optional[str]] = mapped_column(String())
+
+    def __init__(self, **kwargs):
+        for key, value in kwargs.items():
+            setattr(self, key, value)
+
+
+class Label(ArtistOrLabel):
+    __tablename__ = "label"
+
+
+class Artist(ArtistOrLabel):
+    __tablename__ = "artist"
+
+
+# --- Database operation functions ---
+def insert_item(item):
     try:
-        connection = sqlite3.connect(db_file)
-        return connection
-    except sqlite3.Error as e:
-        print(e)
-    return connection
+        db.session.add(item)
+        db.session.commit()
+        return item.id
+    except sqlalchemy.exc.IntegrityError as err:
+        db.session.rollback()
+        print(f'SQLite Integrity Error: \n{err}\n')
+    except Exception as err:
+        db.session.rollback()
+        print(f'Unexpected error: {err}')
 
 
-def create_cursor(connection):
-    try:
-        return connection.cursor()
-    except Exception as e:
-        print(e)
-        return None
-
-
-def create_tables(cur):
-    # Create release table
-    cur.execute("""
-    CREATE TABLE IF NOT EXISTS release(
-    id INTEGER PRIMARY KEY,
-    mbid TEXT UNIQUE,
-    artist_id INTEGER,
-    label_id INTEGER,
-    title TEXT,
-    release_date TEXT,
-    runtime INTEGER,
-    rating INTEGER,
-    listen_date TEXT,
-    track_count INTEGER,
-    country TEXT,
-    genre TEXT,
-    tags TEXT,
-    art TEXT,
-    FOREIGN KEY (artist_id) REFERENCES artist(id)
-        ON UPDATE CASCADE
-        ON DELETE RESTRICT,
-    FOREIGN KEY (label_id) REFERENCES label(id)
-        ON UPDATE CASCADE
-        ON DELETE RESTRICT
+def insert_release(release):
+    new_release = Release(
+        mbid=release.get("mbid"),
+        artist_id=release.get("artist_id", 0),
+        label_id=release.get("label_id", 0),
+        title=release.get("title"),
+        release_year=release.get("release_year", 0),
+        runtime=release.get("runtime", 0),
+        rating=release.get("rating"),
+        listen_date=release.get("listen_date"),
+        track_count=release.get("track_count"),
+        country=release.get("country", 0),
+        art=release.get("art", ''),
+        genre=release.get("genre"),
+        tags=release.get("tags", '')
     )
-    """)
-
-    # Create ARTIST table
-    cur.execute("""
-    CREATE TABLE IF NOT EXISTS artist(
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        mbid TEXT UNIQUE,
-        name TEXT,
-        country TEXT,
-        type TEXT,
-        begin_date TEXT,
-        end_date TEXT,
-        image TEXT
-    )    
-    """)
-
-    # Create LABEL table
-    cur.execute("""
-    CREATE TABLE IF NOT EXISTS label(
-        id INTEGER PRIMARY KEY,
-        mbid TEXT UNIQUE,
-        name TEXT,
-        country TEXT,
-        type TEXT,
-        begin_date TEXT,
-        end_date TEXT,
-        image TEXT
-    )    
-    """)
+    release_id = insert_item(new_release)
+    return release_id
 
 
-def insert_release(cur, con, release):
-    # If we weren't given a label ID, set label ID to 0
-    if "label_id" not in release:
-        release["label_id"] = 0
-
-    query = ("INSERT INTO release (mbid, artist_id, label_id, title, release_date, runtime, rating, listen_date, "
-             "track_count, country, genre, art, tags) VALUES (:mbid, :artist_id, :label_id, :title, :release_date, :runtime, "
-             ":rating, :listen_date, :track_count, :country, :genre, :art, :tags)")
-    try:
-        cur.execute(query, release)
-        con.commit()
-        return True
-    except sqlite3.IntegrityError as e:
-        # print(f"ERROR: {e}")
-        # print("Errored SQLite query:")
-        # print(query)
-        # print("release dictionary:")
-        print(release)
-        print("release already exists in database")
-        return None
+def insert_artist(artist):
+    new_artist = Artist(
+        mbid=artist.get("mbid", "0"),
+        name=artist.get("name"),
+        country=artist.get("country", ""),
+        type=artist.get("type", ""),
+        begin_date=artist.get("begin_date", ""),
+        end_date=artist.get("end_date", ""),
+        image=artist.get("image")
+    )
+    artist_id = insert_item(new_artist)
+    return artist_id
 
 
-def insert_artist(cur, con, artist):
-    query = ("INSERT INTO artist (mbid, name, country, type, begin_date, end_date, image) VALUES"
-             "(:mbid, :name, :country, :type, :begin_date, :end_date, :image)")
-    try:
-        cur.execute(query, artist)
-        con.commit()
-        return cur.lastrowid
-    except sqlite3.IntegrityError as e:
-        print(artist)
-        print("Artist already exists in database")
-        query = "SELECT id FROM artist WHERE mbid = ?"
-        cur.execute(query, (artist['mbid'],))
-        con.commit()
-        return cur.fetchone()[0]
+def insert_label(label):
+    new_label = Label(
+        mbid=label.get("mbid", "0"),
+        name=label.get("name"),
+        country=label.get("country", ""),
+        type=label.get("type", ""),
+        begin_date=label.get("begin_date", ""),
+        end_date=label.get("end_date", ""),
+        image=label.get("image")
+    )
+    label_id = insert_item(new_label)
+    return label_id
 
 
-def insert_label(cur, con, label):
-    query = ("INSERT INTO label (mbid, name, country, type, begin_date, end_date, image) VALUES"
-             "(:mbid, :name, :country, :type, :begin_date, :end_date, :image)")
-    try:
-        cur.execute(query, label)
-        con.commit()
-        return cur.lastrowid
-    except sqlite3.IntegrityError as e:
-        print(label)
-        print("Label already exists in database")
-        query = "SELECT id FROM label WHERE mbid = ?"
-        cur.execute(query, (label['mbid'],))
-        con.commit()
-        return cur.fetchone()[0]
-
-
-def get_stats(cur, con):
-    stats = {}
-
-    cur.execute("SELECT COUNT(*) FROM release")
-    stats["total_listens"] = cur.fetchone()[0]
-
-    cur.execute("SELECT COUNT(*) FROM artist")
-    stats["total_artists"] = cur.fetchone()[0]
-
-    cur.execute("SELECT AVG(rating) FROM release")
-    try:
-        stats["average_rating"] = round(cur.fetchone()[0], 2)
-    except TypeError:
-        stats["average_rating"] = 0
-
-    cur.execute("SELECT AVG(runtime) FROM release")
-    runtime_ms = cur.fetchone()[0]
-    try:
-        stats["average_runtime"] = round((runtime_ms / 60000), 2)
-    except TypeError:
-        stats["average_runtime"] = 0
-
-    cur.execute("SELECT SUM(runtime) FROM release")
-    runtime = cur.fetchone()[0]
-    if runtime is not None:
-        stats["total_runtime"] = round((runtime / 3600000), 2)
-    else:
-        stats["total_runtime"] = 0
-
-    year = str(datetime.datetime.now().year)
-    cur.execute("SELECT COUNT(*) FROM release WHERE SUBSTR(listen_date, 1, 4) = ?",
-                (year,))
-    stats["this_year"] = cur.fetchone()[0]
-
+def get_stats():
+    current_year = str(datetime.datetime.now().year)
     days_this_year = datetime.date.today().timetuple().tm_yday
-    try:
-        stats["per_day"] = round((stats["this_year"] / days_this_year), 2)
-    except ZeroDivisionError:
-        stats["per_day"] = 0
+    # Check if any releases are in the database. If not, skip stats
+    db_length = db.session.query(func.count(Release.id)).scalar()
+    if db_length == 0:
+        return ''
 
-    # most frequent labels
-    most_frequent_labels_query = """
-    SELECT label.name, COUNT(*) as Count 
-    FROM label 
-    JOIN release ON release.label_id = label.id 
-    GROUP BY label.name 
-    ORDER BY Count DESC
-    LIMIT 5;
-    """
-    cur.execute(most_frequent_labels_query)
-    most_frequent_labels = cur.fetchall()
-    stats["most_frequent_labels"] = most_frequent_labels
+    stats = {"total_listens": db.session.query(Release.id).count(),
+             "total_artists": db.session.query(Artist).count(),
+             "total_labels": db.session.query(Label).count(),
+             "average_rating": db.session.query(
+                 func.round(
+                     func.avg(Release.rating), 2)
+             ).scalar(),
+             "average_runtime": round(
+                 (
+                         (
+                             db.session.query(
+                                 func.avg(
+                                     Release.runtime
+                                 )
+                             ).scalar()
+                         ) / 60000
+                 ), 2
+             ),
+             "total_runtime": round(
+                 (
+                         (
+                             db.session.query(
+                                 func.sum(
+                                     Release.runtime
+                                 )
+                             ).scalar()
+                         ) / 3600000
+                 ), 2
+             ),
+             "listens_this_year": db.session.query(func.count(Release.id)).filter(
+                 func.substr(Release.listen_date, 1, 4) == current_year).scalar(),
+             }
+    listens_per_day = stats["listens_this_year"] / days_this_year
+    stats["listens_per_day"] = round(listens_per_day, 2)
 
-    # highest rated labels
-    top_label_query = """
-    SELECT label.name, ROUND(AVG(release.rating), 2) AS Average_Rating, COUNT(release.label_id) as Release_Count 
-    FROM label  
-    JOIN release ON release.label_id = label.id
-    WHERE label.name IS NOT "none" AND label.name IS NOT "[no label]"
-    GROUP BY label.name
-    HAVING Release_Count != 1    
-    ORDER BY Average_Rating DESC
-    LIMIT 5;"""
-    cur.execute(top_label_query)
-    top_labels = cur.fetchall()
-    stats["top_labels"] = top_labels
+    # Top rated labels
+    query = (
+        db.session.query(
+            Label.name,
+            func.round(func.avg(Release.rating), 2).label('average_rating'),
+            func.count(Release.label_id).label('release_count')
+        )
+        .join(Release, Release.label_id == Label.id)
+        .filter(
+            Label.name.notin_(['none', '[no label]']))  # Don't count the entries corresponding to no label for release
+        .group_by(Label.name)
+        .having(func.count(Release.label_id) != 1)  # Don't count labels with only 1 release
+        .order_by(func.round(func.avg(Release.rating), 2).desc())
+        .limit(5)
+        .all()
+    )
+    results = [{'name': row.name,
+                'average_rating': row.average_rating,
+                'release_count': row.release_count}
+               for row in query]
+    stats["favourite_labels"] = results
 
-    # highest rated artists
-    top_artist_query = """
-    SELECT artist.name, ROUND(AVG(release.rating), 2) AS Average_Rating, COUNT(release.artist_id) as Release_Count 
-    FROM artist
-    JOIN release ON release.artist_id = artist.id
-    GROUP BY artist.name
-    HAVING Release_Count != 1    
-    ORDER BY Average_Rating DESC
-    LIMIT 5;"""
-    cur.execute(top_artist_query)
-    top_artists = cur.fetchall()
-    stats["top_artists"] = top_artists
+    # Highest rated artists
+    query = (
+        db.session.query(
+            Artist.name,
+            func.round(func.avg(Release.rating), 2).label('average_rating'),
+            func.count(Release.artist_id).label('release_count')
+        )
+        .join(Release, Release.artist_id == Artist.id)
+        .group_by(Artist.name)
+        .having(func.count(Release.artist_id) != 1)
+        .order_by(func.round(func.avg(Release.rating), 2).desc())
+        .limit(5)
+        .all()
+    )
+    results = [{'name': row.name,
+                'average_rating': row.average_rating,
+                'release_count': row.release_count}
+               for row in query]
+    stats["favourite_artists"] = results
 
-    # most frequent artists
-    frequent_artists_query = """
-    SELECT artist.name, COUNT(release.artist_id) as Count
-    FROM artist
-    JOIN release ON release.artist_id = artist.id
-    WHERE artist.name != "Various Artists"
-    GROUP BY artist.id
-    ORDER BY Count DESC
-    LIMIT 5;
-    """
-    cur.execute(frequent_artists_query)
-    frequent_artists = cur.fetchall()
-    stats["frequent_artists"] = frequent_artists
+    # Most frequent labels
+    query = (
+        db.session.query(
+            Label.name,
+            func.count(Release.label_id).label('count')
+        )
+        .join(Release, Release.label_id == Label.id)
+        .group_by(Label.name)
+        .order_by(func.count(Release.label_id).desc())
+        .limit(5)
+        .all()
+    )
+    results = [{'name': row.name,
+                'count': row.count}
+               for row in query]
+    stats["frequent_labels"] = results
+
+    # Most frequent artists
+    query = (
+        db.session.query(
+            Artist.name,
+            func.count(Release.artist_id).label('count')
+        )
+        .join(Release, Release.artist_id == Artist.id)
+        .group_by(Artist.name)
+        .order_by(func.count(Release.artist_id).desc())
+        .limit(5)
+        .all()
+    )
+    results = [{'name': row.name,
+                'count': row.count}
+               for row in query]
+    stats["frequent_artists"] = results
 
     return stats
 
 
-def update_existing_release(cur, con, release):
-    cur.execute("UPDATE release SET mbid = ?, artist_id = ?, label_id = ?, title = ?, release_date = ?,"
-                "runtime = ?, rating = ?, listen_date = ?, track_count = ?, country = ?, genre = ?, art = ?"
-                "WHERE id = ?",
-                (release["mbid"], release["artist_id"], release["label_id"], release["title"],
-                 release["release_date"], release["runtime"], release["rating"], release["listen_date"],
-                 release["track_count"], release["country"], release["genre"], release["art"], release["id"]))
-    con.commit()
-    print(f"UPDATED {release["title"]}")
+def get_homepage_data():
+    home_data = (
+        db.session.query(
+            Artist.name,
+            Release.title,
+            Release.rating,
+            Release.listen_date,
+            Release.genre,
+            Release.art,
+            Release.mbid,
+            Release.tags
+        )
+        .join(Artist, Artist.id == Release.artist_id)
+        .order_by(Release.listen_date.desc())
+        .limit(10)  # when pagination is implemented this can be raised
+    )
+    return home_data
 
 
-def get_missing_covers(cur, con):
-    cur.execute("SELECT * FROM release WHERE art IS NULL")
-    data = cur.fetchall()
-
-    for release in data:
-        release = list(release)
-        #print(release)
-        mbid = release[1]
-
-        if mbid:
-            art = api.get_cover_art(mbid)
-        else:
-            art = 'https://static.vecteezy.com/system/resources/thumbnails/005/720/408/small_2x/crossed-image-icon-picture-not-available-delete-picture-symbol-free-vector.jpg'
-
-        release[-1] = art
-        release_dict = {"id": release[0], "mbid": release[1], "artist_id": release[2], "label_id": release[3],
-                        "title": release[4], "release_date": release[5], "runtime": release[6],
-                        "rating": release[7], "listen_date": release[8], "track_count": release[9],
-                        "country": release[10], "genre": release[11], "art": release[12]}
-
-        update_existing_release(cur, con, release_dict)
+def get_items(item_type):
+    # Gets all items of a specified type
+    if item_type == 'releases':
+        items = db.session.query(Release).all()
+    elif item_type == 'artists':
+        items = db.session.query(Artist).all()
+    elif item_type == 'labels':
+        items = db.session.query(Label).all()
+    else:
+        raise ValueError('ERROR: Invalid item_type: ', item_type)
+    return items
 
 
-def get_missing_artist_data(cur, con):
-    cur.execute("SELECT * FROM artist")
-    data = cur.fetchall()
+def get_item(item_type, item_id):
+    # Gets a single item specified by its ID
+    if item_type == 'release':
+        release = db.session.query(Release).where(Release.id == item_id)
+        artist = db.session.query(Artist).where(Artist.id == item_id)
+        label = db.session.query(Label).where(Label.id == item_id)
+        item_data = {
+            "release": release[0],
+            "artist": artist[0],
+            "label": label[0]
+        }
+    elif item_type == 'artist':
+        artist = db.session.query(Artist).where(Artist.id == item_id).all()
+        releases = (db.session.query(Release, Label.name)
+                    .join(Label, Release.label_id == Label.id)
+                    .join(Artist, Artist.id == Release.artist_id)
+                    .where(Artist.id == item_id)
+                    .all()
+                    )
+        item_data = {
+            "artist": artist[0],
+            "releases": releases
+        }
+    elif item_type == 'label':
+        label = db.session.query(Label).where(Label.id == item_id).all()
+        releases = (db.session.query(Release, Artist)
+                    .join(Artist, Artist.id == Release.artist_id)
+                    .where(Release.label_id == item_id)
+                    ).all()
+        item_data = {
+            "label": label[0],
+            "releases": releases
+        }
+    else:
+        raise ValueError('ERROR: Invalid item_type: ', item_type)
+    return item_data
 
-    for artist_entry in data:
-        for i, val in enumerate(artist_entry):
-            if not val:
-                new_data = api.get_artist_data(artist_entry[1])
-                query = ("UPDATE artist SET mbid = ?, name = ?, country = ?, type = ?, begin_date = ?, end_date = ?, image = ?"
-                         "WHERE artist.mbid = ?")
-                cur.execute(query, (
-                    new_data["mbid"], new_data["name"], new_data["country"], new_data["type"], new_data["begin_date"],
-                    new_data["end_date"], new_data["image"], new_data["mbid"]
-                    ))
-                con.commit()
-                print(f"UPDATED {new_data['name']}")
-                break
-    return 1
+
+def exists(mbid, item_type):
+    # mbid = the MusicBrainz ID for the entity
+    # item_type = release, artist, or label
+    # Returns a 2 element array:
+    # - 1: True if an item exists in the database, False otherwise
+    # - 2: Item's ID (primary key) in the database if it exists, None otherwise
+    if item_type == 'label':
+        q = db.session.query(Label.id).where(Label.mbid == mbid).scalar()
+    elif item_type == 'artist':
+        q = db.session.query(Artist.id).where(Artist.mbid == mbid).scalar()
+    elif item_type == 'release':
+        q = db.session.query(Release.id).where(Release.mbid == mbid).scalar()
+    else:
+        raise ValueError('Invalid item_type passed')
+    if q:
+        return [True, q]
+    else:
+        return [False, q]
 
 
-def get_missing_label_data(cur, con):
-    cur.execute("SELECT * FROM  label")
-    data = cur.fetchall()
+# ---- INCOMPLETE -----
+def update_release():
+    return 0
 
-    for label_entry in data:
-        if label_entry[1] is None:
-            pass
-        else:
-            for i, val in enumerate(label_entry):
-                if not val:
-                    new_data = api.get_label_data(label_entry[1])
-                    query = ("UPDATE LABEL SET mbid = ?, name = ?, country = ?, type = ?, begin_date = ?, end_date = ?"
-                             "WHERE mbid = ?")
-                    cur.execute(query, (new_data["mbid"], new_data["name"], new_data["country"], new_data["type"],
-                                        new_data["begin_date"], new_data["end_date"],
-                                        new_data["mbid"]))
-                    con.commit()
-                    print(f"UPDATED {new_data['name']}")
-                    break
-    return 1
+
+def update_artist():
+    return 0
+
+
+def update_label():
+    return 0
+
+
+# below should maybe be moved elsewhere
+def get_missing_covers():
+    return 0
+
+
+def get_missing_artist_data():
+    return 0
+
+
+def get_missing_label_data():
+    return 0
