+++ templates/index.html
@@ -53,11 +53,11 @@
             </tr>
             <tr>
                 <td>listens this year</td>
-                <td>{{ stats["this_year"] }}</td>
+                <td>{{ stats["listens_this_year"] }}</td>
             </tr>
             <tr>
                 <td>listens per day (this year)</td>
-                <td>{{ stats["per_day"] }}</td>
+                <td>{{ stats["listens_per_day"] }}</td>
             </tr>
         </table>
     </div>
