+++ templates/label.html
@@ -3,23 +3,19 @@
 {% include 'navbar.html' %}
 <div class="page-content">
     <div>
-    <h1>{{ data[0][2] }}</h1>
-    <img src="{{ data[0][7] }}" style="max-height: 75vh; max-width: 100%;">
+    <h1>{{ data['label'].name }}</h1>
+    <img src="{{ data.label.image }}" style="max-height: 75vh; max-width: 100%;" alt="Label icon">
     </div>
     <div>
         <table class="label-releases-table">
-        {% for release in data[1] %}
+        {% for release in data.releases %}
         <tr>
             <td>
-                <a href="/release/{{ release[1] }}">{{ release[4] }}</a> - {{ release[15] }} ({{ release[5] }})
-            </td>
-            <td>
-                {{ release[5] }}
+                <a href="/release/{{ release[0].id }}">{{ release[0].title }}</a> - {{ release[1].name }} ({{ release[0].release_year }})
             </td>
         </tr>
         {% endfor %}
         </table>
     </div>
 </div>
-<p>{{data[1][0]}}</p>
 </body>
