+++ templates/releases.html
@@ -6,9 +6,9 @@
 <div class="data-grid">
     {% for release in data %}
     <div class="grid-cell">
-        <a href="/release/{{ release[1] }}" class="link">
-        <img src="{{ release[12] }}" style="height:82%; width:100%;">
-        <p>{{ release[4] }} ({{ release[7] }}%)</p>
+        <a href="/release/{{ release.id }}" class="link">
+        <img src="{{ release.art }}" style="height:82%; width:100%;" alt="Release cover art">
+        <p>{{ release.title }} ({{ release.rating }}%)</p>
         </a>
     </div>
     {% endfor %}
