+++ templates/search.html
@@ -82,10 +82,10 @@
                                 </tr>
                                 <tr>
                                     <td>
-                                        <label for="release_date">YEAR</label>
+                                        <label for="release_year">YEAR</label>
                                     </td>
                                     <td>
-                                        <input type="number" min="0" required id="year" name="release_date">
+                                        <input type="number" min="0" required id="year" name="release_year">
                                     </td>
                                 </tr>
                                 <tr>
